requirements.txt
PyYAML==6.0.1
allure-pytest-bdd==2.13.2
pytest==7.4.4
pytest-bdd==7.0.1
selenium==4.16.0


